Revert "solution migration threat analysis response"
