+++ Solutions/ThreatAnalysis&Response/data/Solution_ThreatAnalysis&Response.json
@@ -1,15 +1,14 @@
 {
-  "Name": "Threat Analysis Response",
-  "Author": "Microsoft - support@microsoft.com",
+  "Name": "ThreatAnalysis&Response",
+  "Author": "Sanmit Biraj - v-sabiraj@microsoft.com",
   "Logo": "<img src=\"https://raw.githubusercontent.com/Azure/Azure-Sentinel/master/Workbooks/Images/Logos/Azure_Sentinel.svg\" width=\"75px\" height=\"75px\">",
   "Description": "MITRE ATT&CK® is a globally-accessible knowledge base of adversary tactics and techniques based on real-world observations. The ATT&CK knowledge base is used as a foundation for the development of specific threat models and methodologies in the private sector, in government, and in the cybersecurity product and service community. The MITRE ATT&CK Cloud Matrix provides tactics and techniques representing the MITRE ATT&CK® Matrix for Enterprise covering cloud-based techniques. The Matrix contains information for the following platforms: Azure AD, Office 365, SaaS, IaaS. For more information, see the 💡 [MITRE ATT&CK: Cloud Matrix](https://attack.mitre.org/matrices/enterprise/cloud/)",
+  "WorkbookDescription": "Workbook to showcase MITRE ATT&CK Coverage for Azure Sentinel",
   "Workbooks": [
     "Workbooks/ThreatAnalysis&Response.json",
     "Workbooks/DynamicThreatModeling&Response.json"
   ],
   "Metadata": "SolutionMetadata.json",
-  "BasePath": "C:\\GitHub\\azure\\Solutions\\Threat Analysis Response",
-   "Version": "2.0.0",
-  "TemplateSpec": true,
-  "Is1PConnector": true
+  "BasePath": "C:\\GitHub\\azure\\Solutions\\ThreatAnalysis&Response",
+  "Version": "1.0.14"
 }
